ipcache,watchers,identity: Pass pod UID into ipcache metadata
This commit prevents the deletion of ipcache entries in the case of
StatefulSet restarts. By indexing via the pod UID, we ensure that the
deletion of valid ipcache entries is prevented.

See https://github.com/cilium/cilium/issues/30409 for more details.

+++ pkg/ipcache/ipcache.go
@@ -74,6 +74,8 @@ type K8sMetadata struct {
 	Namespace string
 	// PodName is the Kubernetes pod name behind the IP
 	PodName string
+	// UID is the Kubernetes UID of the pod
+	UID string
 	// NamedPorts is the set of named ports for the pod
 	NamedPorts types.NamedPortMap
 }
@@ -759,11 +761,11 @@ func (ipc *IPCache) GetNamedPorts() (npm types.NamedPortMultiMap) {
 
 // DeleteOnMetadataMatch removes the provided IP to security identity mapping from the IPCache
 // if the metadata cache holds the same "owner" metadata as the triggering pod event.
-func (ipc *IPCache) DeleteOnMetadataMatch(IP string, source source.Source, namespace, name string) (namedPortsChanged bool) {
+func (ipc *IPCache) DeleteOnMetadataMatch(IP string, source source.Source, namespace, name, uid string) (namedPortsChanged bool) {
 	ipc.mutex.Lock()
 	defer ipc.mutex.Unlock()
 	k8sMeta := ipc.getK8sMetadata(IP)
-	if k8sMeta != nil && k8sMeta.Namespace == namespace && k8sMeta.PodName == name {
+	if k8sMeta != nil && k8sMeta.Namespace == namespace && k8sMeta.PodName == name && k8sMeta.UID == uid {
 		return ipc.deleteLocked(IP, source)
 	}
 	return false
@@ -897,6 +899,9 @@ func (m *K8sMetadata) Equal(o *K8sMetadata) bool {
 	if m == o {
 		return true
 	} else if m == nil || o == nil {
+		return false
+	}
+	if m.UID != o.UID {
 		return false
 	}
 	if len(m.NamedPorts) != len(o.NamedPorts) {
